refactor: extract workspace cleanup planning
